+++ src/main/java/br/eng/eliseu/printagent/api/WebPageController.java
@@ -6,5 +6,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 @Controller
 public class WebPageController {
     @GetMapping({"/printAgent", "/printAgent/"})
-    public String painel() { return "forward:/printAgent/index.html"; }
+    public String painel() {
+        return "forward:/printAgent/index.html";
+    }
 }